feat: 중복코드 제거을 위한 BaseSubwayTest 추가
